app/requirements.txt
asgiref==3.7.2
Django==4.2.5
python-dotenv==1.0.0
django-markdownify==0.9.3
django-bootstrap5==23.3
psycopg2-binary==2.9.8
requests==2.31.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.3
gunicorn==21.2.0
Pillow==10.0.1
django-cookie-consent==0.4.0
django-appconf==1.0.5
requests==2.31.0
django-phonenumber-field[phonenumberslite]==7.2.0
